Revert COB heading float-to-short cast fix to match stock engine
GCC at -O2 saturates short(float) to SHORT_MAX (32767) for out-of-range
values, rather than wrapping. The static_cast<short>(static_cast<int>())
fix changes this to wrapping behavior (-15536 for 50000.0f), which
desyncs against demos recorded with the stock engine. Revert to match.

+++ rts/Sim/Units/Scripts/CobInstance.cpp
@@ -237,7 +237,7 @@ void CCobInstance::WindChanged(float heading, float speed)
 {
 	ZoneScoped;
 	Call(COBFN_SetSpeed, int(speed * 3000.0f));
-	Call(COBFN_SetDirection, static_cast<short>(static_cast<int>(heading * RAD2TAANG)));
+	Call(COBFN_SetDirection, short(heading * RAD2TAANG));
 }
 
 
@@ -387,8 +387,8 @@ void CCobInstance::StartBuilding(float heading, float pitch)
 	std::array<int, 1 + MAX_COB_ARGS> callinArgs;
 
 	callinArgs[0] = 2;
-	callinArgs[1] = static_cast<short>(static_cast<int>(heading * RAD2TAANG));
-	callinArgs[2] = static_cast<short>(static_cast<int>(  pitch * RAD2TAANG));
+	callinArgs[1] = short(heading * RAD2TAANG);
+	callinArgs[2] = short(  pitch * RAD2TAANG);
 
 	Call(COBFN_StartBuilding, callinArgs);
 }
@@ -439,8 +439,8 @@ void CCobInstance::AimWeapon(int weaponNum, float heading, float pitch)
 	std::array<int, 1 + MAX_COB_ARGS> callinArgs;
 
 	callinArgs[0] = 2;
-	callinArgs[1] = static_cast<short>(static_cast<int>(heading * RAD2TAANG));
-	callinArgs[2] = static_cast<short>(static_cast<int>(  pitch * RAD2TAANG));
+	callinArgs[1] = short(heading * RAD2TAANG);
+	callinArgs[2] = short(  pitch * RAD2TAANG);
 
 	Call(COBFN_AimPrimary + COBFN_Weapon_Funcs * weaponNum, callinArgs, CBAimWeapon, weaponNum, nullptr);
 }
